added movierating compoenent and changed schema

+++ server/model/rating.js
@@ -1,6 +1,6 @@
 const mongoose = require('mongoose');
 
-const { Schema, Float } = mongoose;
+const { Schema, Decimal128 } = mongoose;
 
 const userRatingSchema = new mongoose.Schema({
     rating: { type: Number, require: true },
@@ -10,7 +10,7 @@ const userRatingSchema = new mongoose.Schema({
 const ratingSchema = new mongoose.Schema({
     movie: { type: Schema.Types.ObjectId, ref: 'Movie' },
     rating: { type: Number, require: true },
-    avgRating: { type: Float, require: true },
+    avgRating: { type: Decimal128, require: true },
     userRatings: [userRatingSchema]
 });
